Code cleanup.

+++ _tests/integration/PicturesCest.php
@@ -270,9 +270,11 @@ class PicturesCest
 
     private function removeDir(string $dir): void
     {
+        if (!is_dir($dir)) {
+            return;
+        }
         $it    = new \RecursiveDirectoryIterator($dir, \RecursiveDirectoryIterator::SKIP_DOTS);
-        $files = new \RecursiveIteratorIterator($it,
-            \RecursiveIteratorIterator::CHILD_FIRST);
+        $files = new \RecursiveIteratorIterator($it, \RecursiveIteratorIterator::CHILD_FIRST);
         foreach ($files as $file) {
             if ($file->isDir()) {
                 rmdir($file->getPathname());
